pipeline/requirements.txt
requests
ipykernel
pandas
python-dotenv
pyodbc
pylint
pytest
